use the path Luke!
We now use a basic path library that scans a directory looking for files
with a particular pattern. We get all matching files and return them to
the user.

Over in uv_socket land we scan /dev/input/by-path and pray to god we
find only 1 device or the user gave us one! If we don't welp we print
all the devices and ask the user to pick one for us to use.

Naturally this needs to be A LOT more intelligent